Discontinuing chr_to_wav maintainance, Implemented windowing and 95-5 ff split

+++ root/integratedDE.py
@@ -45,22 +45,22 @@ Ch= 2
 
 ### MUSIC PARAMETERS
 
-Cl= 300
+Cl= 120
 # Chromosome Length
 # Frames per audio sample
 # This is the number of Time Frames in a single song
 
-Gl= 50
+Gl= 329
 # Gene Length
 # Bins per frame
 # This is the number of frequency Bins in a single Time Frame
 
-Wpb= 160
+Wpb= 24
 # Waves per bin
 # This is the number of waves in a single Bin
 
 
-Frl=0.2
+Frl=0.5
 # Frame Length
 # This is the time in seconds for which a frame lasts
 # IMPORTANT: Frlen*Cl=60 must always hold
@@ -86,12 +86,12 @@ som = SOM()
 # Initialization of Self-Organizing-Map
 
 # Use the below code if there are multiple ffs
-# ff1i=16.5 # Average value of ff1
-# ff2i=36750 # Average value of ff1
-# s1=ff2i/ff1i
-# s2=1
-# w1=0.5 # Weight of ff1
-# w2=0.5 # Weight of ff2
+ff1i=12.5 # Average value of ff1
+ff2i=50 # Average value of ff1
+s1=1
+s2=ff1i/ff2i
+w1=0.95 # Weight of ff1
+w2=0.05 # Weight of ff2
 
 
 def fitnessFunction(Inp):
@@ -108,7 +108,7 @@ def fitnessFunction(Inp):
         float: The fitness value.
     """
 
-    from chr_to_wav import decode
+    from mel_chr_to_wav import decode
     # This is the function that converts chromosomes to wav files.
 
     from driver import computeFitnessValues, compute_SOM_DOB
@@ -136,7 +136,9 @@ def fitnessFunction(Inp):
     # values = dict(computeFitnessValues(rasaNumber=rasaNumber, audioFile=f"gen{generation}-{index}.wav", generation=generation, populationNumber=index))
     # fitnessValue = float(values["fitnessValues"][rasas[rasaNumber-1]]["weightedSum"])
 
-    return fitnessValue**2
+    # return fitnessValue**2
+
+    return w1*s1*(fitnessValue**2)+w2*s2*Q(chromosome)
 
 
 def Q(L):
@@ -149,7 +151,7 @@ def Q(L):
             sum+= abs(L[j][k][0]-L[j][k+1][0])
 
         if j!=Cl-1:
-            sum+=abs(L[j][0][0]-L[j+1][0][0])
+            sum+=abs(L[j][0][0]-L[j+1][k-1][0])
 
     return sum/2000
 
@@ -475,6 +477,10 @@ def main():
 
         if Gn!=Gs: 
             for i in range(Ps):
+
+                if Gn%50==0 and i==Best[0][0]:
+                    continue
+
                 os.remove(f'./audio_output/gen{Gn}-{i}.wav')
                 os.remove(f'./jAudio/gen{Gn}-{i}FV.xml')
                 os.remove(f'./jAudio/gen{Gn}-{i}FK.xml')
